+++ NEWS.rst
@@ -112,6 +112,9 @@ directly to ``BgzfWriter``.
 
 ``Bio.Entrez.local_cache`` can be set to a directory for caching downloaded DTD/XSD files.
 
+``Bio.motifs`` now supports reading PFM from Cys2His2 Zinc Finger Proteins PWM
+Predictor.
+
 As in recent releases, more of our code is now explicitly available under
 either our original "Biopython License Agreement", or the very similar but
 more commonly used "3-Clause BSD License".  See the ``LICENSE.rst`` file for
@@ -123,6 +126,7 @@ possible, especially the following contributors:
 - Anil Tuncel (first contribution)
 - David Cain
 - Fabio Zanini (first contribution)
+- Gert Hulselmans
 - Joao Rodrigues
 - Judith Bernett (first contribution)
 - Luca Monari (first contribution)